Load the zig kernel and jump to it!
THIS IS JUST A HACK! I will hopefully figure out how to find the offset
of the .text section within the kernel elf and then load it to a proper
address. Perhaps somewhere in the higher half?

Also, this code might not work for you because I have manually computed
the offset of .text within the kernel elf and it might change for you
depending on the compiler version etc.