Fix page fault
The offset to the first phdr was incorrectly calculated, which resulted
in the kernel segments not being loaded to their respective vaddrs.

This resulted in jumping to zeroed out memory which the processor
interpreted as an instruction involving dereferencing rax.

The "address" stored in rax at this point wasn't really mapped to any
physical address, hence the page fault.