Setup paging and jump to long mode!
That was fast. I didn't check for the A20 line, nor did I implement any
kind of logging but now that I'm in long mode I can directly start using
a VGA driver maybe(?)

It's not super essential to enable the A20 line for now cause it's
already enabled in qemu. Also tested on bare metal and my laptop seems
to enable it too so cba rn.